Set up 'constants' and 'types' modules
