+++ src/types.rs
@@ -1,0 +1,138 @@
+use candid::{candid_method, CandidType, Decode, Deserialize, Encode, Principal};
+use ic_canister_log::{declare_log_buffer, log};
+use ic_canisters_http_types::{
+    HttpRequest as AssetHttpRequest, HttpResponse as AssetHttpResponse, HttpResponseBuilder,
+};
+use ic_cdk::api::management_canister::http_request::{
+    http_request as make_http_request, CanisterHttpRequestArgument, HttpHeader, HttpMethod,
+    HttpResponse, TransformArgs, TransformContext,
+};
+use ic_cdk::{query, update};
+use ic_nervous_system_common::{serve_logs, serve_logs_v2, serve_metrics};
+use ic_stable_structures::memory_manager::{MemoryId, MemoryManager, VirtualMemory};
+#[cfg(target_arch = "wasm32")]
+use ic_stable_structures::DefaultMemoryImpl;
+use ic_stable_structures::{BoundedStorable, Cell, StableBTreeMap, Storable};
+use std::borrow::Cow;
+use std::cell::RefCell;
+use std::collections::HashMap;
+
+use crate::constants::STRING_STORABLE_MAX_SIZE;
+
+#[derive(Default)]
+pub struct Metrics {
+    pub requests: u64,
+    pub request_cycles_charged: u128,
+    pub request_cycles_refunded: u128,
+    pub request_err_no_permission: u64,
+    pub request_err_service_url_host_not_allowed: u64,
+    pub request_err_http: u64,
+    pub host_requests: HashMap<String, u64>,
+}
+
+// These need to be powers of two so that they can be used as bit fields.
+#[derive(Clone, Debug, PartialEq, CandidType, FromPrimitive, Deserialize)]
+pub enum Auth {
+    Admin = 0b0001,
+    Rpc = 0b0010,
+    RegisterProvider = 0b0100,
+    FreeRpc = 0b1000,
+}
+
+#[derive(Clone, Debug, Default, CandidType, Deserialize)]
+pub struct Metadata {
+    pub nodes_in_subnet: u32,
+    pub next_provider_id: u64,
+    pub open_rpc_access: bool,
+}
+
+#[derive(PartialEq, Eq, PartialOrd, Ord, Clone)]
+pub struct StringStorable(pub String);
+
+#[derive(PartialEq, Eq, PartialOrd, Ord, Clone)]
+pub struct PrincipalStorable(pub Principal);
+
+impl Storable for StringStorable {
+    fn to_bytes(&self) -> std::borrow::Cow<[u8]> {
+        // String already implements `Storable`.
+        self.0.to_bytes()
+    }
+
+    fn from_bytes(bytes: std::borrow::Cow<[u8]>) -> Self {
+        Self(String::from_bytes(bytes))
+    }
+}
+
+impl BoundedStorable for StringStorable {
+    const MAX_SIZE: u32 = STRING_STORABLE_MAX_SIZE;
+    const IS_FIXED_SIZE: bool = false;
+}
+
+impl Storable for PrincipalStorable {
+    fn to_bytes(&self) -> std::borrow::Cow<[u8]> {
+        std::borrow::Cow::from(self.0.as_slice())
+    }
+
+    fn from_bytes(bytes: std::borrow::Cow<[u8]>) -> Self {
+        Self(Principal::from_slice(&bytes))
+    }
+}
+
+impl BoundedStorable for PrincipalStorable {
+    const MAX_SIZE: u32 = 29;
+    const IS_FIXED_SIZE: bool = false;
+}
+
+#[derive(Debug, CandidType)]
+pub struct RegisteredProvider {
+    pub provider_id: u64,
+    pub owner: Principal,
+    pub chain_id: u64,
+    pub service_url: String,
+    pub cycles_per_call: u64,
+    pub cycles_per_message_byte: u64,
+}
+
+#[derive(Debug, CandidType, Deserialize)]
+pub struct RegisterProvider {
+    pub chain_id: u64,
+    pub service_url: String,
+    pub api_key: String,
+    pub cycles_per_call: u64,
+    pub cycles_per_message_byte: u64,
+}
+
+#[derive(Clone, Debug, CandidType, Deserialize)]
+pub struct Provider {
+    pub provider_id: u64,
+    pub owner: Principal,
+    pub chain_id: u64,
+    pub service_url: String,
+    pub api_key: String,
+    pub cycles_per_call: u64,
+    pub cycles_per_message_byte: u64,
+    pub cycles_owed: u128,
+}
+
+impl Storable for Metadata {
+    fn to_bytes(&self) -> std::borrow::Cow<[u8]> {
+        Cow::Owned(Encode!(self).unwrap())
+    }
+    fn from_bytes(bytes: Cow<[u8]>) -> Self {
+        Decode!(&bytes, Self).unwrap()
+    }
+}
+
+impl Storable for Provider {
+    fn to_bytes(&self) -> std::borrow::Cow<[u8]> {
+        Cow::Owned(Encode!(self).unwrap())
+    }
+    fn from_bytes(bytes: Cow<[u8]>) -> Self {
+        Decode!(&bytes, Self).unwrap()
+    }
+}
+
+impl BoundedStorable for Provider {
+    const MAX_SIZE: u32 = 256; // A reasonable limit.
+    const IS_FIXED_SIZE: bool = false;
+}
